feat: split role

+++ studybud/base/views.py
@@ -1,6 +1,8 @@
 from django.shortcuts import render, redirect 
+from django.http import HttpResponse
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
 from .models import Room, Topic
 from .forms import RoomForm
 from django.db.models import Q
@@ -33,7 +35,7 @@ def loginPage(request):
     context = {}
     return render(request, 'base/login_register.html', context)
 
-def logoutUser(request):
+def logoutUser  (request):
     logout(request)
     return redirect('home')
 
@@ -56,6 +58,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -67,10 +70,15 @@ def createRoom(request):
 
     context = {'form': form }
     return render(request, 'base/room_form.html', context)
-    
+
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to edit this room!')
+    
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -80,9 +88,13 @@ def updateRoom(request, pk):
 
     context = {'form': form }
     return render(request, 'base/room_form.html', context) 
-
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to edit this room!')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
